# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: BookList APIs
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /books/{id}:
    # our controller name
    x-swagger-router-controller: book
    get:
      operationId: getOne
      description: get a book
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetBookResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a book
      # define the parameters
      parameters:
        - name: id
          description: Book id
          type: number
          in: path
          required: true
        - name: book
          description: Book properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delBook
      description: delete a book
      # define the parameters
      parameters:
        - name: id
          description: Book id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Book:
    type: object
    properties:
      title:
        type: string
        description: book title
      author:
        type: string
        description: book authoer
      description:
        type: string
        description: book description
    required:
      - title
      - author
...
  GetBookResponse:
    required:
      - id
      - title
      - author
    properties:
      id:
        type: number
      title: 
        type: string
      author:
        type: string
      description:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string